requirements.txt
certifi==2018.11.29
chardet==3.0.4
# crypto==1.4.1
Django==2.1.11 # 为方便测试采用2.1.11版本，在django 2.2版本将sqlite最低支持版本提高为3.8.3，Python3.7以下版本运行会报错
# Django==2.2.4
django-redis==4.10.0
django-rest-framework==0.1.0
# djangorestframework==3.9.0
idna==2.8
Naked==0.1.31
# pycrypto==2.6.1
pytz==2019.2
# PyYAML==3.13
redis==3.0.1
requests==2.21.0
shellescape==3.4.1
# urllib3==1.24.1
python-alipay-sdk==1.9
pyjwt==1.7.1
uWSGI==2.0.17.1
